configured zsh further

+++ zsh/.zshenv
@@ -2,8 +2,10 @@ export ZDOTDIR=${HOME:-~}/.config/dotfiles/zsh
 
 test -f $ZDOTDIR/.zshrc && . $ZDOTDIR/.zshrc
 test -f $ZDOTDIR/.zprofile && . $ZDOTDIR/.zprofile
-test ! -f $HOME/.zshenv && ln -s $ZDOTDIR/.zshenv $HOME/
+if [[ -f $ZDOTDIR/.zshenv ]]; then
+	test ! -f $HOME/.zshenv && ln -s $ZDOTDIR/.zshenv $HOME/
+fi
 
 export EDITOR="/usr/bin/nvim"
 export BROWSER="/mnt/c/Program Files/Firefox Developer Edition/firefox.exe"
-. $HOME/.cargo/env
+test -f $HOME/.cargo/anv && . $HOME/.cargo/env
